refactor(web): single shared dockview layout with per-panel content cache (#456)
Replaces the outer LRU of full dockview instances (DockviewInstanceManager
+ DockviewWorkspaceLayout) with one shared dockview mounted at the app
shell, plus a generic MultiWorkspacePanelHost that owns per-workspace
content caching inside each panel (Chat, Changes, Files, Terminal,
Browser).

Fixes the project-list scroll-jump bug (only one DashboardShell is
mounted now), collapses the duplicated status-watcher subscriptions to
one, and removes the structural-fingerprint eviction dance — layout
edits propagate immediately because the dockview is single.

Architecture
- new: apps/web/src/components/MultiWorkspacePanelHost.tsx — generic
  LRU primitive (size driven by settings.maxCachedWorkspaces). Children
  prop is (workspaceId, wsActive) => ReactNode, rendered inside
  absolutely-positioned divs with opacity toggle. Set-state-during-render
  for first-paint correctness.
- new: apps/web/src/components/SharedDockviewLayout.tsx — successor to
  DockviewWorkspaceLayout. Mounts ONE dockview at the AppShell; owns
  layout persistence (band:dockview-layout-v6 + per-workspace active-tab
  state); cross-panel handlers + per-workspace state (currentFile /
  openFilePath / find-in-file) live here.
- delete: apps/web/src/components/DockviewInstanceManager.tsx
- delete: apps/web/src/components/DockviewWorkspaceLayout.tsx
- delete: apps/web/src/lib/workspace-visibility-store.ts (write-only
  after the refactor — no remaining readers)

Project-list auto-scroll behaviour
- Active workspace card auto-scrolls into view (`block: "center"`) on
  OUTSIDE-the-list navigation: direct URL, browser back/forward, Ctrl+R
  workspace picker.
- IN-LIST activations (mouse click on a card, keyboard Enter on a
  focused card, Enter in the container after arrow nav) suppress the
  auto-scroll via a module-level "recent activation" marker — the
  card is already where the user's cursor / focus is.
- Marker auto-expires after 300ms so it can't suppress a later
  legitimate URL navigation to the same workspace.

Dockview tab restoration
- The per-workspace tab restore effect (apply saved activeState.groups
  on workspace switch) skips no-op setActive() calls: panels that are
  already active in their group, and single-panel groups (the Projects
  edge group). dockview's setActive() has focus side-effects even on
  no-op calls; firing it on the Projects panel was resetting the
  project list scroll and dropping focus on every back-nav to a cached
  workspace.

Testing
- New e2e test: apps/web/e2e/workspace-switch-scroll.spec.ts — three
  Playwright scenarios covering direct-URL centering, in-list-click
  scroll preservation, and browser-back re-centering. Mocks tRPC via
  the existing createTrpcMock helper; 25-project fixture to overflow
  the viewport.
